Update GrizzlyBearEntity.java
Added a rage mode mechanic to the bears. When RevengeGoal is triggered and if the bears health is below 15 (7.5 hearts), they have a 25% chance to gain a massive speed boost which will last until death. It only appears on Hard Difficulty.
I don't know how to add it so it is triggered onHit instead without having to listen to the event. So this have to do for now. 
All that remains is to test this and see if it works.