Fixed targetting, the server does not crashes anymore. And the anger target resets when the entity dies. Nerfed the rage distance (while bear is raging, it will target any player that comes to close until it get's away or dies) by half of it's generic follow range.
Syncing of the rage skin works, fixed angry particles and added one more play of the bear sounds to truly make it clear the rage it is in.
After 30 seconds of not having any player targets, it will return to not raging. Chance of rage will remain 50% during testing but will be set to a default of 25%.